Move zone file parsing and input validation into a submodule
